pyproject.toml
[tool.poetry]
name = "kline-scraping"
version = "0.1.0"
description = ""
authors = ["raychang <raychang@xrex.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.11.12"
tqdm = "^4.67.1"
pandas = "^2.2.3"
motor = "^3.7.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
